Add a kerberos security module.
