+++ frameworks/hdfs/tests/auth.py
@@ -5,8 +5,8 @@ import itertools
 import logging
 import base64
 
+from security import kerberos
 
-import sdk_cmd
 import sdk_hosts
 
 
@@ -26,9 +26,8 @@ def get_service_principals(service_name: str, realm: str) -> list:
     :return: A list of said principals
     """
     primaries = ["hdfs", "HTTP"]
-    fqdn = sdk_hosts.autoip_host(service_name, "")
 
-    instances = [
+    tasks = [
         "name-0-node",
         "name-0-zkfc",
         "name-1-node",
@@ -42,25 +41,15 @@ def get_service_principals(service_name: str, realm: str) -> list:
     ]
 
     principals = []
-    for (primary, instance) in itertools.product(primaries, instances):
-        principals.append(
-            "{primary}/{instance}{fqdn}@{REALM}".format(
-                primary=primary,
-                instance=instance,
-                fqdn=fqdn,
-                REALM=realm
-            )
-        )
+    for (primary, task) in itertools.product(primaries, tasks):
+        instance = sdk_hosts.autoip_host(service_name, task)
+        principals.append(kerberos.genererate_principal(primary, instance, realm))
 
     for primary in USERS:
-        principals.append(
-            "{primary}@{REALM}".format(
-                primary=primary,
-                REALM=realm
-            )
-        )
+        principals.append(kerberos.genererate_principal(primary, realm=realm, instance=None))
 
-    http_principal = "HTTP/api.{}.marathon.l4lb.thisdcos.directory".format(service_name)
+    http_instance = "api.{}.marathon.l4lb.thisdcos.directory".format(service_name)
+    http_principal = kerberos.genererate_principal("HTTP", http_instance, realm)
 
     principals.append(http_principal)
     return principals
@@ -81,27 +70,3 @@ def get_principal_to_user_mapping() -> str:
         rules.append("RULE:[1:$1@$0](^{user}@.*$)s/.*/{user}/".format(user=user))
 
     return base64.b64encode('\n'.join(rules).encode("utf-8")).decode("utf-8")
-
-
-def write_krb5_config_file(task: str, filename: str, krb5: object) -> str:
-    """
-    Generate a Kerberos config file.
-    TODO(elezar): This duplicates functionality in frameworks/kafka/tests/auth.py
-    TODO(elezar): Move to testing/security/kerberos.py
-    """
-    output_file = filename
-
-    log.info("Generating %s", output_file)
-    krb5_file_contents = ['[libdefaults]',
-                          'default_realm = {}'.format(krb5.get_realm()),
-                          '',
-                          '[realms]',
-                          '  {realm} = {{'.format(realm=krb5.get_realm()),
-                          '    kdc = {}'.format(krb5.get_kdc_address()),
-                          '  }', ]
-    log.info("%s", krb5_file_contents)
-
-    output = sdk_cmd.create_task_text_file(task, output_file, krb5_file_contents)
-    log.info(output)
-
-    return output_file
