+++ testing/security/__init__.py
@@ -1,4 +1,5 @@
+from . import kerberos
 from . import transport_encryption
 
 
-__all__ = [transport_encryption]
+__all__ = [kerberos, transport_encryption]
